align login page to use the same logic
